% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{RblFiles}
\alias{RblFiles}
\title{List files available at Bloomberg}
\usage{
RblFiles()
}
\value{
Vector of character strings representing the filenames available at Bloomberg
}
\description{
Retrieve files available at Bloomberg.
}
\examples{
\dontrun{
# Run RblConnect first
RblFiles()
}

}
