% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{RblUrl}
\alias{RblUrl}
\title{Get Bloomberg connection Url}
\usage{
RblUrl()
}
\value{
Url string
}
\description{
Retrieve the Bloomberg connection Url.
}
\examples{
\dontrun{
# Run RblConnect first
RblUrl()
}

}
